Add method to get joint index including universe
The member index() method is only usable if the joint is not NULL.
We could not get index from an empty pointer that supposedly represents
the "universe" joint. Having a static method that deals with this will
help to streamline a lot of codes that deal with joint index. Previously
this method is only available in hpp-core::RelativeMotion which is not
obvious at all.